Widen SoundId to 32 bits
House sounds are numbered type + 100 * (roomSoundId + 300), which runs up to 39999 in MM7, and dsounds.bin
stores ids as uint32. An int16_t SoundId wrapped every house sound from room 28 up to a negative value. That
never broke a lookup, because the formula and the loader wrapped the same way and no two MM7 ids collide
after wrapping, but it made those ids impossible to name in the enum. Nothing depends on the width: the save
and table snapshots convert through their own uint16_t and uint32_t fields.

+++ src/Engine/Data/SoundEnums.h
@@ -10,7 +10,7 @@
  * There's 2000+ sounds in MM7, and there is little point in adding all of them here via codegen. Only the ids that are
  * actually used in the codebase should be kept in this enum.
  */
-enum class SoundId : int16_t {
+enum class SoundId : int32_t {
     SOUND_Invalid = 0,
     SOUND_enter = 6,
     SOUND_WoodDoorClosing = 7,
